Add missing save() call at the end of open()
If changes in self.info are done, self.save() writes them into
self.oldinfo.
When modifying/creating the object self.oldinfo and self.info are
compared and the changes are written to LDAP. As there no save() call is
done the change can't be detected.

Probably open() (and therefore save()) was called mutliple times, so that this never occurred as bug
and this change is not a regression.

open() was e.g. called in openSuperordinate.